Move protobuf generated code to io.spine.web.firebase package.
