+++ firebase-web/src/main/java/io/spine/web/firebase/query/FirebaseQueryResult.java
@@ -20,7 +20,7 @@
 
 package io.spine.web.firebase.query;
 
-import io.spine.web.firebase.client.NodePath;
+import io.spine.web.firebase.NodePath;
 import io.spine.web.query.QueryProcessingResult;
 
 import javax.servlet.ServletResponse;
