+++ firebase-web/src/main/java/io/spine/web/firebase/subscription/FirebaseSubscriptionBridge.java
@@ -30,8 +30,8 @@ import io.spine.client.SubscriptionVBuilder;
 import io.spine.client.Topic;
 import io.spine.client.grpc.QueryServiceGrpc;
 import io.spine.web.firebase.FirebaseClient;
+import io.spine.web.firebase.NodePath;
 import io.spine.web.firebase.NodePaths;
-import io.spine.web.firebase.client.NodePath;
 import io.spine.web.firebase.query.QueryNodePath;
 import io.spine.web.query.service.AsyncQueryService;
 import io.spine.web.subscription.SubscriptionBridge;
