+++ firebase-web/src/main/java/io/spine/web/firebase/subscription/SubscriptionRecord.java
@@ -25,8 +25,8 @@ import io.spine.client.QueryResponse;
 import io.spine.json.Json;
 import io.spine.protobuf.AnyPacker;
 import io.spine.web.firebase.FirebaseClient;
+import io.spine.web.firebase.NodePath;
 import io.spine.web.firebase.NodeValue;
-import io.spine.web.firebase.client.NodePath;
 import io.spine.web.firebase.subscription.diff.Diff;
 import io.spine.web.firebase.subscription.diff.DiffCalculator;
