+++ firebase-web/src/main/proto/spine/web/firebase/client.proto
@@ -24,7 +24,7 @@ package spine.web.firebase;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.web.firebase.client";
+option java_package = "io.spine.web.firebase";
 option java_multiple_files = true;
 option java_outer_classname = "FirebaseClientProto";
