+++ firebase-web/src/test/java/io/spine/web/firebase/rest/RestClientTest.java
@@ -23,11 +23,11 @@ package io.spine.web.firebase.rest;
 import com.google.api.client.http.ByteArrayContent;
 import com.google.api.client.http.GenericUrl;
 import com.google.common.testing.NullPointerTester;
+import io.spine.web.firebase.DatabaseUrl;
 import io.spine.web.firebase.DatabaseUrls;
+import io.spine.web.firebase.NodePath;
 import io.spine.web.firebase.NodePaths;
 import io.spine.web.firebase.NodeValue;
-import io.spine.web.firebase.client.DatabaseUrl;
-import io.spine.web.firebase.client.NodePath;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
